requirements.txt
requests>=2.32.0
supabase>=2.18.0
python-dotenv>=1.1.0
